M - atulaizar

+++ src/main/java/br/com/ithappens/model/scantech/FechamentoDiarioScantechDTO.java
@@ -17,7 +17,7 @@ public class FechamentoDiarioScantechDTO {
     private Long idFilial;
     private Integer pdv;
     private BigDecimal montoVentaLiquida;
-    private Integer montoCancelaciones;
+    private BigDecimal montoCancelaciones;
     private Integer cantidadMovimientos;
     private Integer cantidadCancelaciones;
